Extract common conversion utils for img
